requirements_base.txt
otree==2.5.8
numpy>=1.19.1
otreeutils==0.9.2
pandas>=1.1.0
python-dateutil>=2.8.1
django_radiogrid==1.1.0